requirements.txt
Django>=6.0.1
djangorestframework>=3.14.0
django-filter>=23.0
djangorestframework-simplejwt>=5.3.0
